+++ backend/src/main/java/com/art/drinkgpt/repositories/DrinkRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DrinkRepository extends JpaRepository<Drink, Long> {
+public interface DrinkRepository extends JpaRepository<Drink, Integer> {
     @Query("SELECT DISTINCT d FROM Drink d JOIN d.drinkIngredients di WHERE di.ingredient.id = :ingredientId")
-    Page<Drink> findByIngredientId(Long ingredientId, Pageable pageable);
+    Page<Drink> findByIngredientId(Integer ingredientId, Pageable pageable);
 }
