+++ backend/src/main/java/com/art/drinkgpt/services/DrinkService.java
@@ -5,9 +5,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface DrinkService {
-    Page<DrinkDTO> getAllDrinks(Pageable pageable, Long ingredientId);
-    DrinkDTO getDrinkById(Long id);
+    Page<DrinkDTO> getAllDrinks(Pageable pageable, Integer ingredientId);
+    DrinkDTO getDrinkById(Integer id);
     DrinkDTO createDrink(DrinkDTO request);
-    DrinkDTO updateDrink(Long id, DrinkDTO request);
-    void deleteDrink(Long id);
+    DrinkDTO updateDrink(Integer id, DrinkDTO request);
+    void deleteDrink(Integer id);
 }
